GitHub contributions count added to news index page
The number of contributions made to larvalabs/pullup now gets displayed below the GitHub account link on the news/index page.

Added the node-cache dependency to packages.json. Used for caching GitHub API responses.

Added the octonode node module dependency to packages.json. Used to simplify GitHub API requests.

Added the 'pullup' object to res.locals. This object gets passed to the client-side JS by means of data.jade and globals.js. Since our content security policy disallows the use of inline scripts, template variables values cannot be set directly to JS variables. To get around this, properties of the pullup object are inserted into the DOM as data-* attributes, which are then retrieved and stored as globals in globals.js.

Added the components directory for custom server-side modules.

Added the module constants.js and with it a DEBUG flag.